+++ src/content/headsup/index.tsx
@@ -1,17 +1,25 @@
-import { Button, Icon, Intent } from "@blueprintjs/core";
-import { IconNames } from "@blueprintjs/icons";
-import { l } from "@r2c/extension/analytics";
 import DomElementLoadedWatcher from "@r2c/extension/content/github/DomElementLoadedWatcher";
-import { ErrorHeadsUp, LoadingHeadsUp, UnsupportedHeadsUp } from "@r2c/extension/content/headsup/NonIdealHeadsup";
-import { PreflightChecklist, PreflightChecklistFetch, PreflightChecklistItemType } from "@r2c/extension/content/headsup/PreflightChecklist";
+import {
+  ErrorHeadsUp,
+  LoadingHeadsUp,
+  UnsupportedHeadsUp
+} from "@r2c/extension/content/headsup/NonIdealHeadsup";
+import {
+  PreflightChecklist,
+  PreflightChecklistFetch,
+  PreflightChecklistItemType
+} from "@r2c/extension/content/headsup/PreflightChecklist";
 import UsedBy from "@r2c/extension/content/headsup/UsedBy";
+import LastUpdatedBadge from "@r2c/extension/content/LastUpdatedBadge";
 import RepoPackageSection from "@r2c/extension/content/PackageCopyBox";
 import { R2CLogo } from "@r2c/extension/icons";
+import { ExtractedRepoSlug } from "@r2c/extension/utils";
 import * as React from "react";
 import * as ReactDOM from "react-dom";
 import "./index.css";
 
 interface HeadsUpProps {
+  repoSlug: ExtractedRepoSlug;
   onChecklistItemClick(
     itemType: PreflightChecklistItemType
   ): React.MouseEventHandler<HTMLElement>;
@@ -22,15 +30,7 @@ interface HeadsupState {
 }
 
 class NormalHeadsUp extends React.PureComponent<HeadsUpProps, HeadsupState> {
-  public state: HeadsupState = {
-    closed: false
-  };
-
   public render() {
-    if (this.state.closed) {
-      return null;
-    }
-
     return (
       <PreflightChecklistFetch>
         {({ loading, error, data, response }) => (
@@ -39,17 +39,23 @@ class NormalHeadsUp extends React.PureComponent<HeadsUpProps, HeadsupState> {
             {response != null &&
               response.repo.status === 404 && <UnsupportedHeadsUp />}
             {error &&
-              (response == null ||
-              response.repo.status !== 404) && <ErrorHeadsUp error={error} />}
+              (response == null || response.repo.status !== 404) && (
+                <ErrorHeadsUp error={error} />
+              )}
             {data && (
               <div className="r2c-repo-headsup checklist-headsup">
                 <header>
-                  <h1>Checklist</h1>
-                  <Button
-                    icon={IconNames.SMALL_CROSS}
-                    minimal={true}
-                    onClick={l("preflight-closed", this.closeMessage)}
-                  />
+                  <div className="checklist-left">
+                    <span className="preflight-logo">preflight</span>
+                  </div>
+                  <div className="checklist-right">
+                    <LastUpdatedBadge
+                      commitHash={data.findings.commitHash}
+                      lastUpdatedDate={new Date(data.repo.analyzedAt)}
+                      repoSlug={this.props.repoSlug}
+                    />
+                    <R2CLogo />
+                  </div>
                 </header>
                 <div className="repo-headsup-body">
                   <div className="repo-headsup-checklist">
@@ -63,32 +69,6 @@ class NormalHeadsUp extends React.PureComponent<HeadsUpProps, HeadsupState> {
                     <UsedBy pkg={data.pkg} />
                   </div>
                 </div>
-                <footer>
-                  <div className="preflight-legend">
-                    <span className="legend-check legend-entry">
-                      <Icon
-                        icon={IconNames.SMALL_TICK}
-                        intent={Intent.SUCCESS}
-                      />{" "}
-                      Good
-                    </span>
-                    <span className="legend-warn legend-entry">
-                      <Icon
-                        icon={IconNames.SYMBOL_TRIANGLE_UP}
-                        intent={Intent.WARNING}
-                      />{" "}
-                      Careful
-                    </span>
-                    <span className="legend-missing legend-entry">
-                      <Icon icon={IconNames.MINUS} /> N/A
-                    </span>
-                  </div>
-                  <div className="preflight-footer-side">
-                    <div className="preflight-logo">
-                      preflight <R2CLogo />
-                    </div>
-                  </div>
-                </footer>
               </div>
             )}
           </>
@@ -96,10 +76,6 @@ class NormalHeadsUp extends React.PureComponent<HeadsUpProps, HeadsupState> {
       </PreflightChecklistFetch>
     );
   }
-
-  private closeMessage: React.MouseEventHandler<HTMLElement> = e => {
-    this.setState({ closed: true });
-  };
 }
 
 interface RepoHeadsUpState {